+++ harmony_regridding_service/regridding_service.py
@@ -23,7 +23,7 @@ from harmony_regridding_service.file_io import (
     clone_variables,
     transfer_metadata,
 )
-from harmony_regridding_service.grid import compute_target_area
+from harmony_regridding_service.grid import compute_target_areas
 from harmony_regridding_service.resample import (
     cache_resamplers,
     copy_resampled_dimension_variables,
@@ -58,12 +58,12 @@ def regrid(
     )
 
     try:
-        target_area = compute_target_area(message, input_filepath, var_info)
+        target_areas = compute_target_areas(message, input_filepath, var_info)
     except InvalidCRSResampling as e:
         logger.warning(f'{e}: Returning your input file unchanged.')
         return input_filepath
 
-    resampler_cache = cache_resamplers(input_filepath, var_info, target_area)
+    resampler_cache = cache_resamplers(input_filepath, var_info, target_areas)
 
     target_filepath = generate_output_filename(input_filepath, is_regridded=True)
 
@@ -72,9 +72,9 @@ def regrid(
         Dataset(target_filepath, mode='w', format='NETCDF4') as target_ds,
     ):
         transfer_metadata(source_ds, target_ds)
-        transfer_resampled_dimensions(source_ds, target_ds, target_area, var_info)
+        transfer_resampled_dimensions(source_ds, target_ds, target_areas, var_info)
         crs_map = write_grid_mappings(
-            target_ds, get_resampled_dimension_pairs(var_info), target_area
+            target_ds, get_resampled_dimension_pairs(var_info), target_areas
         )
 
         vars_to_process = var_info.get_all_variables()
@@ -86,7 +86,7 @@ def regrid(
         vars_to_process -= cloned_vars
 
         dimension_vars = copy_resampled_dimension_variables(
-            source_ds, target_ds, target_area, var_info
+            source_ds, target_ds, target_areas, var_info
         )
         logger.info(f'processed dimension variables: {dimension_vars}')
         vars_to_process -= dimension_vars